% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_sele.R
\name{Filter}
\alias{Filter}
\title{Filter allele_ages by quality of tree}
\usage{
Filter(allele_ages, qual)
}
\arguments{
\item{allele_ages}{data.table. Obtained from get.allele_ages()}

\item{qual}{data.frame. qual file}
}
\value{
Returns a data table.
}
\description{
This function filters out selection pvalues on bad trees using summary statistics on tree quality stored in qual.
}
\examples{
Filter(allele_ages, qual)

}
