% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_haps.R
\name{get_hap_matrix}
\alias{get_hap_matrix}
\title{Get haplotype matrix (L x N)}
\usage{
get_hap_matrix(haps)
}
\arguments{
\item{haps}{data.table.}
}
\value{
Returns a matrix.
}
\description{
This function extracts the haplotype matrix from .haps data frames
}
\examples{
haps <- read.haps(system.file("extdata/example.haps.gz", package = "relater"))
get_hap_matrix(haps)
}
