% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygenic_test.R
\name{PolyTest_Init}
\alias{PolyTest_Init}
\title{Initialise polygenic selection test}
\usage{
PolyTest_Init(allele_ages)
}
\arguments{
\item{allele_ages}{data.table. Obtained from get.allele_ages()}
}
\value{
Returns a list containing a data frame (quantiles) and a list of data tables (allele_ages_quantiles).
}
\description{
This function initialises the polygenic selection test.
}
\examples{
PolyTest_Init(allele_ages)
}
