% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygenic_test.R
\name{PolyTest_ResampleSNPs}
\alias{PolyTest_ResampleSNPs}
\title{Resample selection p-values}
\usage{
PolyTest_ResampleSNPs(DAF, quantiles, allele_ages_quantiles)
}
\arguments{
\item{DAF}{int. Derived allele frequency}

\item{quantiles}{data frame. Obtained from PolyTest_Init.}

\item{allele_ages_quantiles}{List of data tables. Obtained from PolyTest_Init.}
}
\value{
Returns a numeric 1x20 matrix.
}
\description{
Given DAF, resample 20 selection p-values of SNPs selected uniformly at random with matching DAF
}
\examples{
PolyTest_ResampleSNPs(DAF, quantiles, allele_ages_quantiles)
}
