% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics_parse.R
\name{read.avg_rate}
\alias{read.avg_rate}
\title{Parse average rate file}
\usage{
read.avg_rate(filename)
}
\arguments{
\item{filename}{string. Filename}
}
\value{
Returns a data frame.
}
\description{
This function parses a *_avg.rate file, which stores average mutation rates inferred by Relate
}
\examples{
read.avg_rate("example_avg.rate")
}
