% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics_parse.R
\name{read.coal}
\alias{read.coal}
\title{Parse coal file}
\usage{
read.coal(filename, ...)
}
\arguments{
\item{filename}{string. Filename}
}
\value{
Returns a data frame.
}
\description{
This function parses a .coal file, which stores coalescence rates inferred by Relate
}
\examples{
read.coal("example.coal")
}
