% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics_parse.R
\name{read.freq}
\alias{read.freq}
\title{Parse freq file}
\usage{
read.freq(filename, CHR = NA)
}
\arguments{
\item{filename}{string. Filename}

\item{CHR}{int. Chromosome index.}
}
\value{
Returns a data frame.
}
\description{
This function parses a .freq file, which stores frequencies in Relate-inferred genealogies
}
\examples{
read.freq("example.freq")
}
