% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics_parse.R
\name{read.haps}
\alias{read.haps}
\title{Parse haps file}
\usage{
read.haps(filename, ...)
}
\arguments{
\item{filename}{string. Filename}

\item{...}{Any other parameters for fread}
}
\value{
Returns a data table.
}
\description{
This function uses fread to parse a .haps file, which is part of the haps/sample output file format of SHAPEIT2.
}
\examples{
read.haps("example.haps")

}
