% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics_parse.R
\name{read.lin}
\alias{read.lin}
\title{Parse lin file}
\usage{
read.lin(filename, CHR = NA)
}
\arguments{
\item{filename}{string. Filename}

\item{CHR}{int. Chromosome index.}
}
\value{
Returns a data frame.
}
\description{
This function parses a .lin file, which stores number of lineages remaining in Relate-inferred genealogies
}
\examples{
read.lin(system.file("extdata/example.lin.gz", package = "relater"))
}
