% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics_parse.R
\name{read.mut}
\alias{read.mut}
\title{Parse mut file}
\usage{
read.mut(filename, CHR = NA, ...)
}
\arguments{
\item{filename}{string. Filename}

\item{CHR}{int. Chromosome index.}

\item{...}{Any other parameters for fread}
}
\value{
Returns a data table.
}
\description{
This function uses fread to parse a .mut file, which is part of the anc/mut output file format of SHAPEIT2.
}
\examples{
read.haps("example.mut")

}
