% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics_parse.R
\name{read.qual}
\alias{read.qual}
\title{Parse qual file}
\usage{
read.qual(filename, CHR = NA)
}
\arguments{
\item{filename}{string. Filename}

\item{CHR}{int. Chromosome index.}
}
\value{
Returns a data frame.
}
\description{
This function parses a .qual file, which stores summary statistics on the quality of Relate-inferred genealogies
}
\examples{
read.qual("example.qual")
}
