% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics_parse.R
\name{read.sample}
\alias{read.sample}
\title{Parse sample file}
\usage{
read.sample(filename, ...)
}
\arguments{
\item{filename}{string. Filename}

\item{...}{Any other parameters for fread}
}
\value{
Returns a data table.
}
\description{
This function uses fread to parse a .sample file, which is part of the haps/sample output file format of SHAPEIT2.
}
\examples{
read.haps("example.sample")

}
