% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics_parse.R
\name{read.sele}
\alias{read.sele}
\title{Parse sele file}
\usage{
read.sele(filename, CHR = NA)
}
\arguments{
\item{filename}{string. Filename}

\item{CHR}{int. Chromosome index.}
}
\value{
Returns a data frame.
}
\description{
This function parses a .sele file, which stores selection p-values inferred by Relate
}
\examples{
read.sele("example.sele")

}
