% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_sele.R
\name{filter.allele_ages}
\alias{filter.allele_ages}
\title{Filter allele_ages by quality of tree}
\usage{
filter.allele_ages(allele_ages, qual)
}
\arguments{
\item{allele_ages}{data.table. Obtained from get.allele_ages()}

\item{qual}{data.frame. qual file}
}
\value{
Returns a data table.
}
\description{
This function filters out selection pvalues on bad trees using summary statistics on tree quality stored in qual.
}
\examples{
# get allele ages
mut         <- read.mut(system.file("extdata/example.mut.gz", package = "relater"), CHR = 1)
sele        <- read.sele(system.file("extdata/example.sele.gz", package = "relater"), CHR = 1)
freq        <- read.freq(system.file("extdata/example.freq.gz", package = "relater"), CHR = 1)
allele_ages <- get.allele_ages(mut, freq, sele)

# read quality file
qual        <- read.qual(system.file("extdata/example.qual.gz", package = "relater"), CHR = 1)
filter.allele_ages(allele_ages, qual)
}
