% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_sele.R
\name{get.allele_ages}
\alias{get.allele_ages}
\title{Compile allele ages file}
\usage{
get.allele_ages(mut, freq, sele)
}
\arguments{
\item{mut}{data.table. mut file}

\item{freq}{data.frame. freq file}

\item{sele}{data.frame. sele file}
}
\value{
Returns a data table.
}
\description{
This function aggregates mut, freq, and sele files inferred by Relate into an allele_ages file.
}
\examples{
get.allele_ages(mut, freq, sele)

}
